SpringBoot2로 Rest api 만들기(8) –  SpringSecurity를 이용한 인증 및 권한부여

+++ src/main/java/com/rest/api/repo/UserJpaRepo.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface UserJpaRepo extends JpaRepository<User, Integer> {
+public interface UserJpaRepo extends JpaRepository<User, Long> {
 
     Optional<User> findByUid(String email);
 }
